Move shared curated types into a new common_types module
DeploymentMetadata, SensorIdentity, SensorExtrinsics, PlatformIdentity,
PlatformSensor, VesselIdentity, and VesselSensor are no longer
descriptor-specific now that the deployment bundle design (#199) reuses
them too, and types.py is being phased out. Pure move; no behavior change.
