+++ src/afft/deployment/descriptor_types.py
@@ -4,142 +4,13 @@ from datetime import datetime
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from .types import DeploymentMetadata
-
-
-class SensorIdentity(BaseModel):
-    """
-    Curated identity of a sensor, filled by enrichment from the deployment
-    catalog. Vendor and product are empty where the hardware has no published
-    record.
-
-    Attributes
-    ----------
-    label: Human-readable sensor name.
-    vendor: Manufacturer (e.g. ``"Teledyne RDI"``).
-    product: Product name (e.g. ``"Work Horse Navigator"``).
-    type: Sensor type (e.g. ``"dvl"``).
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    label: str
-    vendor: str
-    product: str
-    type: str
-
-
-class SensorExtrinsics(BaseModel):
-    """
-    A sensor's mounting pose, in the reference frame of the body it is mounted
-    on — stated by the descriptor section the sensor lives in. Filled by
-    enrichment.
-
-    Attributes
-    ----------
-    locx: X translation in metres.
-    locy: Y translation in metres.
-    locz: Z translation in metres.
-    rotx: Roll angle in radians.
-    roty: Pitch angle in radians.
-    rotz: Yaw angle in radians.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    locx: float
-    locy: float
-    locz: float
-    rotx: float
-    roty: float
-    rotz: float
-
-
-class PlatformSensor(BaseModel):
-    """
-    A sensor mounted on the deployment's platform. The whole roster is curated
-    — enrichment writes it from the assigned catalog profile — so a platform
-    section either has no sensors at all or has them fully identified.
-
-    Attributes
-    ----------
-    key: Curated sensor identifier (e.g. ``"dvl_teledyne_navigator"``); the
-        catalog lookup key.
-    message_topics: RAW telemetry topics the sensor emits, carried from the
-        catalog so the descriptor states the mapping a build used. Empty where
-        the sensor emits no topic.
-    identity: Curated sensor metadata; ``None`` only on a roster no enrichment
-        wrote.
-    extrinsics: Mounting pose in the vehicle (SNAME) body frame; ``None``
-        where the sensor has no surveyed pose.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    key: str
-    message_topics: list[str] = Field(default_factory=list)
-    identity: SensorIdentity | None = None
-    extrinsics: SensorExtrinsics | None = None
-
-
-class VesselSensor(BaseModel):
-    """
-    A sensor mounted on the deployment's support vessel, its pose in the ship
-    reference frame rather than the vehicle (SNAME) body frame.
-
-    Attributes
-    ----------
-    key: Curated sensor identifier (e.g. ``"usbl_linkquest_transceiver"``);
-        the catalog lookup key.
-    message_topics: RAW telemetry topics the sensor emits; empty for the
-        topside sensors, whose data arrives through a file role instead.
-    identity: Curated sensor metadata; ``None`` only on a roster no enrichment
-        wrote.
-    extrinsics: Mounting pose in the ship reference frame; ``None`` where the
-        sensor has no surveyed pose.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    key: str
-    message_topics: list[str] = Field(default_factory=list)
-    identity: SensorIdentity | None = None
-    extrinsics: SensorExtrinsics | None = None
-
-
-class PlatformIdentity(BaseModel):
-    """
-    Curated identity of the deployment's platform, filled by enrichment. Only
-    ``platform_class`` has a counterpart in the deployment data files, as
-    ``system.vehicle_name``.
-
-    Attributes
-    ----------
-    platform_label: Human-readable platform name (e.g. ``"AUV Sirius"``).
-    platform_class: Vehicle class / system config vehicle name (e.g.
-        ``"SEABED"``).
-    platform_operator: Operating institution (e.g. ``"ACFR"``).
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    platform_label: str
-    platform_class: str
-    platform_operator: str
-
-
-class VesselIdentity(BaseModel):
-    """
-    Curated identity of the deployment's support vessel.
-
-    Attributes
-    ----------
-    vessel_name: Support vessel name (e.g. ``"RV Linnaeus"``).
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    vessel_name: str
+from .common_types import (
+    DeploymentMetadata,
+    PlatformIdentity,
+    PlatformSensor,
+    VesselIdentity,
+    VesselSensor,
+)
 
 
 class DeploymentPlatformSection(BaseModel):
